Separate logic that handles event listners; move get image method to uitilities folder
