+++ src/ui/weather.js
@@ -1,6 +1,6 @@
-import getTemperature from '../utils/temperature';
+import convertTemperature from '../events/convert_temp';
+import getImage from '../utils/image';
 
-const getImage = (name) => `./src/assets/${name}.jpg`;
 
 const render = (data, city, countryName) => {
 
@@ -64,62 +64,8 @@ const render = (data, city, countryName) => {
   }
 
   const UIChKFahrenheit = document.querySelector('.chk-fahrenheit');
-  UIChKFahrenheit.addEventListener('change', (e) => {
-    const UIMinMax = document.querySelector('.min-max');
-    const UITemperature = document.querySelector('.temperature');
-    const UIMin = document.querySelector('.min');
-    const UIMax = document.querySelector('.max');
-    console.log(UITemperature, UIMin, UIMax);
-
-    const current = UITemperature.textContent;
-    const currentMin = UIMin.textContent;
-    const currentMax = UIMax.textContent;
-    console.log(current, currentMin, currentMax);
-
-    if (e.target.checked) {
-      const { temp, min, max } = getTemperature(
-        current,
-        currentMin,
-        currentMax,
-        'F'
-      );
-
-      UITemperature.innerHTML = `
-        <span>${temp}</span>
-        <span>&deg;F</span>
-        `;
-
-      UIMinMax.innerHTML = `
-              <div className="min-max">
-                Min: <span class="min">${min}</span>
-                <span>&deg;C</span>
-                Max: <span class="max">${max}</span>
-                <span>&deg;F</span>
-              </div>
-        `;
-    } else {
-      const { temp, min, max } = getTemperature(
-        current,
-        currentMin,
-        currentMax,
-        'C'
-      );
-
-      UITemperature.innerHTML = `
-          <span>${temp}</span>
-        <span>&deg;C</span>
-          `;
-
-      UIMinMax.innerHTML = `
-              <div className="min-max">
-                Min: <span class="min">${min}</span>
-                <span>&deg;C</span>
-                Max: <span class="max">${max}</span>
-                <span>&deg;C</span>
-              </div>
-        `;
-    }
-  });
+  UIChKFahrenheit.addEventListener('change', e => convertTemperature(e))
+  
 };
