+++ src/utils/image.js
@@ -1,0 +1,3 @@
+const getImage = (name) => `./src/assets/${name}.jpg`;
+
+export default getImage;
